Provide a test config
In a mocha test it's common for a test to look like this:

```
it('returns OK (200)', () => {
  return request(app).get(url).expect(200);
});
```

With `arrow-body-style` and `newline-per-chained-call` on this would
force the test to look like this:

```
it('returns OK (200)', () => request(app)
  .get(url)
  .expect(200));
```

In the case of mocha tests the convention is more readable. So a new
test config is provided that disables these rules.

Use it by in your `test/.eslintrc.json`:

```
{
  "extends": "@hmcts/eslint-config/test"
}
```