Allow magic numbers and undefined in tests
In a test suite it's desired to test against edge cases and bad usage.
In these tests you would use a variety of numbers and `undefined` to see
what your unit does under those values.

This relaxes the linting for tests to allow test suites to perform these
checks without twisting the code.